\name{end.view}
\alias{end.view}
\title{end view}

\usage{\method{end}{view}()}


\references{\url{https://www.texmacs.org}, \url{https://www.texmacs.org/tmweb/plugins/statistics.en.html}}
\author{Michael Lachmann Tamarlin}



\seealso{\code{\link{v}},\code{\link{start.view}},\code{\link{postscript}}}
\examples{
start.view()
plot(1:10,10:1)
v()
end.view()
}
