\name{t.cat.in2out}
\alias{t.cat.in2out}
\title{t cat in2out}

\usage{\method{t}{cat.in2out}(filename, before = paste(DATA_BEGIN, "ps: ", sep = ""), 
    after = paste("\\n", DATA_END, "\\n"), fix.html.bugs = F)}
\arguments{
  \item{filename}{
}
  \item{before}{
}
  \item{after}{
}
  \item{fix.html.bugs}{
}
}



\author{Michael Lachmann Tamarlin}





