\name{t.display.html.help}
\alias{t.display.html.help}
\title{Display R help within the R session in TeXmacs}
\description{This function will display a help document within the current R session.
Usually, you will not call this function, but instead do
\code{assign("?",t.display.html.help)}}
\usage{\method{t}{display.html.help}(e1, package = NULL, lib.loc = NULL, verbose = getOption("verbose"), 
    try.all.packages = getOption("help.try.all.packages"), help_type = getOption("help_type"))}
\arguments{
  \item{e1}{
}
  \item{package}{
}
  \item{lib.loc}{
}
  \item{verbose}{
}
  \item{try.all.packages}{
}
  \item{help_type}{
}
}



\author{Michael Lachmann Tamarlin}





