\name{t.tab.comp}
\alias{t.tab.comp}
\title{Send tab completion to TeXmacs}
\description{t.tab.comp is usually not invoked by the user. When TeXmacs asks for a tab completion,
this function is invoked, which sends the completion back to TeXmacs.}
\usage{\method{t}{tab.comp}(s, curs)}
\arguments{
  \item{s}{
}
  \item{curs}{
}
}



\author{Michael Lachmann Tamarlin}





