\name{TeXmacs-package}
\alias{TeXmacs-package}
\docType{package}
\title{R interface with TeXmacs}
\description{This package interfaces with TeXmacs - inserting figures, help files and more.}
\details{
\tabular{ll}{Package: \tab TeXmacs\cr
Title: \tab R interface with TeXmacs\cr
Version: \tab 0.15\cr
Author: \tab Michael Lachmann Tamarlin\cr
Maintainer: \tab Michael Lachmann Tamarlin <mlachmann@gmail.com>\cr
License: \tab GPL\cr}
}
\author{Michael Lachmann Tamarlin}

\keyword{ package }


