\name{translator_texmacs}
\alias{translator_texmacs}
\title{translator texmacs}

\usage{translator_texmacs(x, size = c("normalsize", "tiny", "scriptsize", 
    "footnotesize", "small", "large", "Large", "LARGE", "huge", 
    "Huge"))}
\arguments{
  \item{x}{
}
  \item{size}{
}
}



\author{Michael Lachmann Tamarlin}





